Add a new API endpoint at `/api/hello/` that returns a JSON response with the message "Hello World" when the query parameter `key` is present. If the `key` query parameter is not present, return a JSON response with the message "key query parameter is required" and a status code of 501.
This commit adds the necessary code changes to the `api/views.py` file to implement the new `get_hello` function. It also includes corresponding tests in the `api/tests.py` file to ensure the endpoint behaves as expected.
