+++ copilot/api/urls.py
@@ -5,4 +5,5 @@ from . import views
 
 urlpatterns = [
     path('time/', views.get_current_time),
+    path('hello/', views.get_hello)
 ]
